Title: 사칙연산 함수, Time: 107ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 241015/사칙연산 함수/quadratic-operations-function.py
@@ -4,7 +4,7 @@ def cal(a, o, c):
     elif o == '-':
         return a - c
     elif o == '/':
-        return a / c
+        return int(a / c)
     else:
         return a * c
